requirements.txt
anyio==3.7.1
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Django==4.1.10
django-allauth==0.54.0
django-cors-headers==4.1.0
django-dotenv==1.4.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.4.0
drf-writable-nested==0.7.0
drf-yasg==1.21.5
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pymongo==4.4.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
sniffio==1.3.0
sqlparse==0.2.4
uritemplate==4.1.1
urllib3==2.0.3
